ajout de la couche dto et couche service

+++ src/main/java/fr/scriptoria/repositories/mySqlRepositories/ProjectRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import fr.scriptoria.models.entities.ProjectEntity;
 
-public interface ProjectRepository extends JpaRepository<ProjectEntity, Long> {
+public interface ProjectRepository extends JpaRepository<ProjectEntity, Integer> {
     
 }
